% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_algos.R
\name{algo_compare}
\alias{algo_compare}
\title{Compare Algorithms}
\usage{
algo_compare(algos = NULL, outcome = NULL, budget = 1000, revenue = 10,
  compare_with = "full", MC = 0.25, int_price = 2)
}
\arguments{
\item{algos}{algorithm input as a matrix with each column corresponding to one matrix.}

\item{outcome}{outcome that the algorithms predict should be passsed in as a vector of the same length as the matrix.}

\item{budget}{amount in dollars that the client has to spend}

\item{revenue}{amount that the client will gain for each successful conversion (outcome==1)}

\item{compare_with}{Should the comparison model be a "full" model or a "null" model with only an intercept?}

\item{MC}{marginal cost of producing an additional unit for each outcome (constant value as a percentage of revenue)}

\item{int_price}{The minimum amount that a customer purchases given a successful outcome
Function to evaluate algorithm input. 
The function produces a list that includes the different LOO (leave-one-out predictive density) values for each
of the algorithm inputs and a comparable estimate based on the budget that looks at net revenues.}
}
