% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_algos.R
\name{algo_pay}
\alias{algo_pay}
\title{Payment Calculation for Algorithms
This function produces estimates of how much to pay each algorithm based on the overall budget and available profits.}
\usage{
algo_pay(obj = NULL)
}
\arguments{
\item{obj}{The list object from the `algo_compare` function
The function produces a data frame with calculations for each algorithm}
}
