% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_algos.R
\name{algo_plot}
\alias{algo_plot}
\title{Plot Algorithms
Function that generates plots showing either LOO (leave-one-out predictive density) for all algorithms or profit/loss figures for each of the algorithms.}
\usage{
algo_plot(obj = NULL, plot_type = "cash")
}
\arguments{
\item{plot_type}{"loo" for predictive density, "cash" for budget/profit differences}

\item{input}{A list as produced by the `algo_compare` function.}
}
