% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_algos.R
\name{algo_sim}
\alias{algo_sim}
\title{Simulate Algorithms
Function that will produce example algorithms based on simulated data. The simulation assumes that the algorithms can be ranked on their average predictive performance for a binomial (0/1) distributed outcome. Coefficient values are on the logit scale.}
\usage{
algo_sim(N = 1000, algo_num = 3, algo_coefs = seq(0, 0.5, length.out =
  algo_num), algo_errs = rep(0.25, algo_num), intercept = -1)
}
\arguments{
\item{N}{Number of observations to simulate for each algorithm}

\item{algo_num}{Number of algorithms to simulate}

\item{algo_coefs}{Vector of coefficient values for algorithms on the logit scale. If absent, the algorithms will be placed on an even scale from 0 to .25.}

\item{algo_errs}{Vector of SDs for each of the algorithms. Set at 0.25 if empty.}

\item{intercept}{A numeric value reflecting the average outcome (on the logit scale)
Function produces a list with an outcome and a matrix of predictors for use in the `compare_algos` function.}
}
